distribute manual test cases
